refactor: seperate basic info component in signup page
