+++ src/pages/SignUp/components/index.ts
@@ -1,1 +1,2 @@
 export { PasswordInput } from './PasswordInput';
+export { BasicInfo } from './BasicInfo';
